pyproject.toml
[project]
name = "file-ops"
version = "0.2.0"
description = "A local file operations skill: convert, inspect, archive, and extract text from files"
authors = [
    { name = "wowyuarm", email = "wuyuang1007@gmail.com" }
]
dependencies = [
    "pillow",
    "pandas",
    "pdf2docx",
    "pdfkit",
    "markdown",
    "openpyxl>=3.1.5",
    "pymupdf",
    "python-docx",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build]
only-packages = true
